feat: unwrap global namespace to flat exports + barrel
